Extract records lifecycle CLI
